+++ src/bot/constants/state.py
@@ -1,0 +1,1 @@
+CHECK: int = 1
