--- src/bot/utils.py
@@ -1,4 +1,0 @@
-STICKER_ID: str = (
-    "CAACAgIAAxkBAAIDwGUEItYpuU5DXDRigDA1M9SL-AdcAALTJAACHRAwSj-yFNWnj6hDMAQ"
-)
-CHECK: int = 1
